rename: `sample` directory to `utility`
