+++ src/utility/Sample.cpp
@@ -1,4 +1,4 @@
-#include "sample/Sample.h"
+#include "utility/Sample.h"
 
 float breq = 0;
